Clean up validate
